pa4/requirements.txt
asn1crypto==0.23.0
cffi==1.11.2
cryptography==2.1.3
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
pycparser==2.18
pycrypto==2.6.1
pyDes==2.0.1
pyOpenSSL==17.3.0
six==1.11.0
